backend/analysis/requirements.txt
fastapi
uvicorn
kafka-python
pydantic
sqlalchemy
psycopg2-binary
pandas
aiokafka